feat: refactor category API to server actions
