+++ src/modules/todos/edit-todo.page.tsx
@@ -3,8 +3,8 @@ import Link from "next/link";
 import { notFound } from "next/navigation";
 import { Button } from "@/components/ui/button";
 import { requireAuth } from "@/modules/auth/utils/auth-utils";
+import { getAllCategories } from "@/modules/todos/actions/get-categories.action";
 import { getTodoById } from "@/modules/todos/actions/get-todo-by-id.action";
-import { getAllCategories } from "@/modules/todos/server/categories.server";
 import { TodoForm } from "./components/todo-form";
 import todosRoutes from "./todos.route";
